mv pool nomination status fns to hook
