+++ packages/app/src/contexts/Pools/BondedPools/types.ts
@@ -2,15 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import type { Dispatch, SetStateAction } from 'react'
-import type {
-  AnyJson,
-  BondedPool,
-  MaybeAddress,
-  NominationStatus,
-  NominationStatuses,
-  Nominator,
-  PoolTab,
-} from 'types'
+import type { AnyJson, BondedPool, Nominator, PoolTab } from 'types'
 
 export interface BondedPoolsContextState {
   queryBondedPool: (poolId: number) => Promise<BondedPool | undefined>
@@ -18,11 +10,6 @@ export interface BondedPoolsContextState {
   updateBondedPools: (bondedPools: BondedPool[]) => void
   addToBondedPools: (bondedPool: BondedPool) => void
   removeFromBondedPools: (poolId: number) => void
-  getPoolNominationStatus: (
-    nominator: MaybeAddress,
-    address: MaybeAddress
-  ) => NominationStatus
-  getPoolNominationStatusCode: (statuses: NominationStatuses | null) => string
   replacePoolRoles: (poolId: number, roleEdits: AnyJson) => void
   poolSearchFilter: (pools: BondedPool[], searchTerm: string) => BondedPool[]
   bondedPools: BondedPool[]
