+++ packages/app/src/hooks/useNominationStatus/index.tsx
@@ -4,11 +4,17 @@
 import { useBalances } from 'contexts/Balances'
 import { useEraStakers } from 'contexts/EraStakers'
 import { useActivePool } from 'contexts/Pools/ActivePool'
+import { useBondedPools } from 'contexts/Pools/BondedPools'
 import { useStaking } from 'contexts/Staking'
 import { useValidators } from 'contexts/Validators/ValidatorEntries'
 import { useSyncing } from 'hooks/useSyncing'
 import { useTranslation } from 'react-i18next'
-import type { BondFor, MaybeAddress, NominationStatus } from 'types'
+import type {
+  BondFor,
+  MaybeAddress,
+  NominationStatus,
+  NominationStatuses,
+} from 'types'
 
 export const useNominationStatus = () => {
   const { t } = useTranslation()
@@ -17,6 +23,7 @@ export const useNominationStatus = () => {
   const { getNominations } = useBalances()
   const { getValidators } = useValidators()
   const { activePoolNominations } = useActivePool()
+  const { bondedPools, poolsNominations } = useBondedPools()
   const { syncing, activePoolSynced, accountSynced } = useSyncing([
     'era-stakers',
   ])
@@ -141,10 +148,56 @@ export const useNominationStatus = () => {
     }
   }
 
+  // Get bonded pool nomination statuses
+  const getPoolNominationStatus = (
+    nominator: MaybeAddress,
+    nomination: MaybeAddress
+  ): NominationStatus => {
+    const pool = bondedPools.find((p) => p.addresses.stash === nominator)
+
+    if (!pool) {
+      return 'waiting'
+    }
+
+    // get pool targets from nominations metadata
+    const nominations = poolsNominations[pool.id]
+    const targets = nominations ? nominations.targets : []
+    const target = targets.find((item) => item === nomination)
+
+    if (!target) {
+      return 'waiting'
+    }
+
+    const nominationStatus = getNominationsStatusFromTargets(nominator, [
+      target,
+    ])
+    return getPoolNominationStatusCode(nominationStatus)
+  }
+
+  // Determine bonded pool's current nomination statuse
+  const getPoolNominationStatusCode = (statuses: NominationStatuses | null) => {
+    let status: NominationStatus = 'waiting'
+
+    if (statuses) {
+      for (const childStatus of Object.values(statuses)) {
+        if (childStatus === 'active') {
+          status = 'active'
+          break
+        }
+        if (childStatus === 'inactive') {
+          status = 'inactive'
+        }
+      }
+    }
+    return status
+  }
+
   return {
     getNominationStatus,
     getNominationSetStatus,
     nominationStatusSyncing,
     getNominationsStatusFromTargets,
+    getPoolNominationStatus,
+    getPoolNominationStatusCode,
   }
 }
